refactor: modify modules structure
